requirements.txt
deepface
flask
pyyaml
easydict